pyproject.toml
[project]
name = "cdc-pdf-pipeline"
version = "0.1.0"
description = "CDC event-driven PDF pipeline demo"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "taskiq>=0.11",
    "taskiq-redis>=0.5",
    "google-cloud-pubsub>=2.26",
    "google-cloud-storage>=2.19",
    "pypdf>=5.0",
    "pydantic-settings>=2.7",
    "typer>=0.15",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "structlog>=24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "mypy>=1.13",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cdc_pdf_pipeline"]
